refactor(domain): one implementation of "what would you score here"
The personalized Score tier list needs the same estimate PUMBILITY makes, and
could not call it: PlayerProgress already references ChartIntelligence, so a
contract query the other way closes a cycle. Both can see Domain, and
TierListProcessor is already there for exactly this reason.

What moved is more than the arithmetic. CohortEstimator was always pure and
always shareable; the part that would actually have drifted is the gathering
around it — who counts as a peer, and what level each of them held at the
moment they set the score they are being read for. That is now ScoreProjector,
and PumbilityProjectionSaga sheds three ports along with it.

The competitive window stays a parameter with no default. PUMBILITY quotes the
projected number, so it wants the window measured most accurate; a tier list
only ranks a folder against itself. Making callers say which they are is what
stops one silently inheriting the other's answer.

Behaviour-preserving: the saga's own suite passes with the real projector
wired in, constructor aside.
