+++ ScoreTracker/ScoreTracker.Tests/DomainTests/CohortEstimatorTests.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using ScoreTracker.PlayerProgress.Domain;
+using ScoreTracker.Domain.Services;
 using Xunit;
 
 namespace ScoreTracker.Tests.DomainTests;
